Break into snare and server
